requirements.txt
fastapi
gunicorn
uvicorn
pyodbc
azure-appconfiguration-provider
